remove old aggregates code
