+++ code/go/0chain.net/smartcontract/dbs/event/blobber_aggregate.go
@@ -1,13 +1,8 @@
 package event
 
 import (
-	"math"
-
-	"0chain.net/chaincore/config"
 	"0chain.net/smartcontract/dbs/model"
 	"github.com/0chain/common/core/currency"
-	"github.com/0chain/common/core/logging"
-	"go.uber.org/zap"
 )
 
 type BlobberAggregate struct {
@@ -34,203 +29,6 @@ type BlobberAggregate struct {
 	InactiveRounds      int64         `json:"InactiveRounds"`
 	RankMetric          float64       `json:"rank_metric"`
 	Downtime            uint64        `json:"downtime"`
-}
-
-func (edb *EventDb) updateBlobberAggregate(round, pageAmount int64, gs *Snapshot) {
-	currentBucket := round % config.Configuration().ChainConfig.DbSettings().AggregatePeriod
-
-	exec := edb.Store.Get().Exec("CREATE TEMP TABLE IF NOT EXISTS temp_ids "+
-		"ON COMMIT DROP AS SELECT id as id FROM blobbers where bucket_id = ?",
-		currentBucket)
-	if exec.Error != nil {
-		logging.Logger.Error("error creating temp table", zap.Error(exec.Error))
-		return
-	}
-
-	exec = edb.Store.Get().Exec("CREATE TEMP TABLE IF NOT EXISTS old_temp_ids "+
-		"ON COMMIT DROP AS SELECT blobber_id as id FROM blobber_snapshots where bucket_id = ?",
-		currentBucket)
-	if exec.Error != nil {
-		logging.Logger.Error("error creating old temp table", zap.Error(exec.Error))
-		return
-	}
-
-	var count int64
-	r := edb.Store.Get().Raw("SELECT count(*) FROM temp_ids").Scan(&count)
-	if r.Error != nil {
-		logging.Logger.Error("getting ids count", zap.Error(r.Error))
-		return
-	}
-	if count == 0 {
-		return
-	}
-	pageCount := count / edb.PageLimit()
-
-	logging.Logger.Debug("blobber aggregate/snapshot started", zap.Int64("round", round), zap.Int64("bucket_id", currentBucket), zap.Int64("page_limit", edb.PageLimit()))
-	for i := int64(0); i <= pageCount; i++ {
-		edb.calculateBlobberAggregate(gs, round, edb.PageLimit(), i*edb.PageLimit())
-	}
-
-}
-
-// paginate divides `count` of items in exactly `round` pages and returns
-// the size of the page, current page number and amount of subpages if needed
-// for example, we have round=101, pageAmount=2, count=11, then
-// size will be 6, current page 1, and subpage count 1
-func paginate(round, pageAmount, count, pageLimit int64) (int64, int64, int) {
-	size := int64(math.Ceil(float64(count) / float64(pageAmount)))
-	currentPageNumber := round % pageAmount
-
-	subpageCount := 1
-	if size > pageLimit {
-		subpageCount = int(math.Ceil(float64(size) / float64(pageLimit)))
-	}
-	return size, currentPageNumber, subpageCount
-}
-
-func (edb *EventDb) calculateBlobberAggregate(gs *Snapshot, round, limit, offset int64) {
-	var ids []string
-	r := edb.Store.Get().
-		Raw("select id from temp_ids ORDER BY ID limit ? offset ?", limit, offset).Scan(&ids)
-	if r.Error != nil {
-		logging.Logger.Error("getting ids", zap.Error(r.Error))
-		return
-	}
-
-	var currentBlobbers []*Blobber
-	result := edb.Store.Get().Model(&Blobber{}).
-		Where("blobbers.id in (select id from temp_ids ORDER BY ID limit ? offset ?)", limit, offset).
-		Joins("Rewards").
-		Find(&currentBlobbers)
-
-	if result.Error != nil {
-		logging.Logger.Error("getting current blobbers", zap.Error(result.Error))
-		return
-	}
-
-	oldBlobbers, err := edb.getBlobberSnapshots(limit, offset)
-	if err != nil {
-		logging.Logger.Error("getting blobber snapshots", zap.Error(err))
-		return
-	}
-
-	var (
-		oldBlobbersProcessingMap = MakeProcessingMap(oldBlobbers)
-		aggregates               []BlobberAggregate
-		gsDiff                   Snapshot
-		old                      BlobberSnapshot
-		ok                       bool
-	)
-
-	for _, current := range currentBlobbers {
-		if processingEntity, found := oldBlobbersProcessingMap[current.ID]; !found {
-			old = BlobberSnapshot{ /* zero values */ }
-			gsDiff.BlobberCount += 1
-		} else {
-			old, ok = processingEntity.Entity.(BlobberSnapshot)
-			if !ok {
-				logging.Logger.Error("error converting processable entity to blobber snapshot")
-				continue
-			}
-		}
-
-		// Case: blobber becomes killed/shutdown
-		if current.IsOffline() && !old.IsOffline() {
-			handleOfflineBlobber(&gsDiff, old)
-			continue
-		}
-
-		aggregate := BlobberAggregate{
-			Round:     round,
-			BlobberID: current.ID,
-			BucketID:  current.BucketId,
-		}
-		aggregate.WritePrice = (old.WritePrice + current.WritePrice) / 2
-		aggregate.Capacity = (old.Capacity + current.Capacity) / 2
-		aggregate.Allocated = (old.Allocated + current.Allocated) / 2
-		aggregate.SavedData = (old.SavedData + current.SavedData) / 2
-		aggregate.ReadData = (old.ReadData + current.ReadData) / 2
-		aggregate.TotalStake = (old.TotalStake + current.TotalStake) / 2
-		aggregate.TotalRewards = (old.TotalRewards + current.Rewards.TotalRewards) / 2
-		aggregate.OffersTotal = (old.OffersTotal + current.OffersTotal) / 2
-		aggregate.OpenChallenges = (old.OpenChallenges + current.OpenChallenges) / 2
-		aggregate.TotalBlockRewards = (old.TotalBlockRewards + current.TotalBlockRewards) / 2
-		aggregate.TotalStorageIncome = (old.TotalStorageIncome + current.TotalStorageIncome) / 2
-		aggregate.TotalReadIncome = (old.TotalReadIncome + current.TotalReadIncome) / 2
-		aggregate.TotalSlashedStake = (old.TotalSlashedStake + current.TotalSlashedStake) / 2
-		aggregate.Downtime = current.Downtime
-
-		aggregate.ChallengesPassed = current.ChallengesPassed
-		aggregate.ChallengesCompleted = current.ChallengesCompleted
-
-		if current.ChallengesCompleted == 0 {
-			aggregate.RankMetric = 0
-		} else {
-			aggregate.RankMetric = float64(current.ChallengesPassed) / float64(current.ChallengesCompleted)
-		}
-		aggregates = append(aggregates, aggregate)
-
-		gsDiff.SuccessfulChallenges += int64(current.ChallengesPassed - old.ChallengesPassed)
-		gsDiff.TotalChallenges += int64(current.ChallengesCompleted - old.ChallengesCompleted)
-		gsDiff.TotalStaked += int64(current.TotalStake - old.TotalStake)
-		gsDiff.StorageTokenStake += int64(current.TotalStake - old.TotalStake)
-		gsDiff.AllocatedStorage += current.Allocated - old.Allocated
-		gsDiff.MaxCapacityStorage += current.Capacity - old.Capacity
-		gsDiff.UsedStorage += current.SavedData - old.SavedData
-		gsDiff.TotalRewards += int64(current.Rewards.TotalRewards - old.TotalRewards)
-		gsDiff.BlobberTotalRewards += int64(current.Rewards.TotalRewards - old.TotalRewards)
-
-		// Change in staked storage (staked_storage = total_stake / write_price)
-		oldSS := old.Capacity
-		if old.WritePrice > 0 {
-			oldSS = int64((float64(old.TotalStake) / float64(old.WritePrice)) * GB)
-		}
-		newSS := current.Capacity
-		if current.WritePrice > 0 {
-			newSS = int64((float64(current.TotalStake) / float64(current.WritePrice)) * GB)
-		}
-		gsDiff.StakedStorage += (newSS - oldSS)
-	}
-
-	gs.ApplyDiff(&gsDiff)
-
-	if len(aggregates) > 0 {
-		if result := edb.Store.Get().Create(&aggregates); result.Error != nil {
-			logging.Logger.Error("saving aggregates", zap.Error(result.Error))
-		}
-	}
-
-	if len(currentBlobbers) > 0 {
-		if err := edb.addBlobberSnapshot(currentBlobbers, round); err != nil {
-			logging.Logger.Error("saving blobbers snapshots", zap.Error(err))
-		}
-	}
-
-	logging.Logger.Debug("blobber aggregate/snapshots finished successfully",
-		zap.Int("current_blobbers", len(currentBlobbers)),
-		zap.Int("old_blobbers", len(oldBlobbers)),
-		zap.Int("aggregates", len(aggregates)),
-		zap.Any("global_snapshot_after", gs),
-	)
-}
-
-func handleOfflineBlobber(gsDiff *Snapshot, old BlobberSnapshot) {
-	gsDiff.SuccessfulChallenges += int64(-old.ChallengesPassed)
-	gsDiff.TotalChallenges += int64(-old.ChallengesCompleted)
-	gsDiff.AllocatedStorage += -old.Allocated
-	gsDiff.MaxCapacityStorage += -old.Capacity
-	gsDiff.UsedStorage += -old.SavedData
-	gsDiff.TotalRewards += int64(-old.TotalRewards)
-	gsDiff.TotalStaked += int64(-old.TotalStake)
-	gsDiff.StorageTokenStake += int64(-old.TotalStake)
-	gsDiff.BlobberCount -= 1
-
-	if old.WritePrice > 0 {
-		ss := int64((float64(old.TotalStake) / float64(old.WritePrice)) * GB)
-		gsDiff.StakedStorage += -ss
-	} else {
-		gsDiff.StakedStorage += -old.Capacity
-	}
 }
 
 func (edb *EventDb) CreateBlobberAggregates(blobbers []*Blobber, round int64) error {
