+++ code/go/0chain.net/smartcontract/dbs/event/blobber_test.go
@@ -13,6 +13,7 @@ import (
 
 	"github.com/0chain/common/core/currency"
 	"github.com/0chain/common/core/logging"
+	"github.com/go-faker/faker/v4"
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
@@ -217,4 +218,20 @@ func setUpBlobbers(t *testing.T, eventDb *EventDb, number int) []string {
 	}
 	require.NoError(t, eventDb.addBlobbers(blobbers))
 	return ids
+}
+
+func buildMockBlobber(t *testing.T, pid string, bucket int64) Blobber {
+	var curBlobber Blobber
+	err := faker.FakeData(&curBlobber)
+	require.NoError(t, err)
+	curBlobber.ID = pid
+	curBlobber.DelegateWallet = OwnerId
+	curBlobber.BucketId = bucket
+	curBlobber.BaseURL = fmt.Sprintf("http://url-%v.com", pid)
+	curBlobber.WritePrice += 10
+	curBlobber.Capacity += int64(curBlobber.TotalStake) * int64(GB)
+	curBlobber.IsKilled = false
+	curBlobber.IsShutdown = false
+	curBlobber.Rewards = ProviderRewards{}
+	return curBlobber
 }
